sprint 2

+++ TP_CLINICA/src/app/solicitar-turno/solicitar-turno.component.ts
@@ -5,6 +5,9 @@ import { HorariosService } from '../services/horarios.service';
 import { FirestoreService } from '../services/firestore.service';
 import { ToastService } from '../services/toast.service';
 import { SwalService } from '../services/swal.service';
+import { UsuarioPipe } from '../shared/pipes/usuario.pipe';
+import { SortByNameAndLastNamePipe } from '../shared/pipes/sort-by-name-and-last-name.pipe';
+import { Perfil } from '../models/enums/perfil';
 
 @Component({
   selector: 'app-solicitar-turno',
@@ -12,7 +15,8 @@ import { SwalService } from '../services/swal.service';
   styleUrls: ['./solicitar-turno.component.css']
 })
 export class SolicitarTurnoComponent {
-
+  tituloBuscador: string = 'Ingrese Nombre | Apellido | Dni de paciente';
+  filaSeleccionada: any;
   cargando: boolean = false;
   spinner: boolean = false;
   spinnerEspecialistas: boolean = false;
@@ -22,10 +26,13 @@ export class SolicitarTurnoComponent {
   turnos: any;
   suscripcionEspecialidades: any;
   suscripcionEspecialistas: any;
+  suscripcionUsuarios: any;
   suscripcionTurnos: any;
   especialidadSeleccionada: any;
   usuarioActual: any;
-
+  pacienteSeleccionado: any = null;
+  usuarioPipe: any;
+  ordenarPorNombreYApellidoPipe: any;
   constructor(
     private usuarioService: UsuarioService,
     private especialidadService: EspecialidadService,
@@ -38,10 +45,15 @@ export class SolicitarTurnoComponent {
     this.eliminarSuscripcionEspecialidades();
     this.eliminarSuscripcionEspecialistas();
     this.eliminarSuscripcionTurnos();
+    this.eliminarSuscripcionUsuarios();
   }
   eliminarSuscripcionEspecialistas() {
     if (this.suscripcionEspecialistas)
       this.suscripcionEspecialistas.unsubscribe();
+  }
+
+  esAdmin() {
+    return this.usuarioActual?.perfil == Perfil.administrador;
   }
 
   eliminarSuscripcionEspecialidades() {
@@ -53,25 +65,47 @@ export class SolicitarTurnoComponent {
     if (this.suscripcionTurnos)
       this.suscripcionTurnos.unsubscribe();
   }
+  eliminarSuscripcionUsuarios() {
+    if (this.suscripcionUsuarios)
+      this.suscripcionUsuarios.unsubscribe();
+  }
 
   ngOnDestroy() {
     this.eliminarSuscripciones();
   }
 
   ngOnInit() {
+    this.cargando = true;
+    this.usuarioPipe = new UsuarioPipe();
+    this.ordenarPorNombreYApellidoPipe = new SortByNameAndLastNamePipe();
+
     this.traerEspecialidades();
     this.usuarioService.obtenerUsuarioActual().then(x => {
       this.usuarioActual = x;
     });
+
+    this.cargando = false;
   }
 
   seleccionEspecialista(especialista: any) {
     this.spinnerTurnos = true;
     this.eliminarSuscripcionTurnos();
     this.suscripcionTurnos = this.horariosService.traerListaDeItemsFiltradoPor_TRES_CamposDe_IGUALDAD('idMedico', especialista.id, 'idEspecialidad', this.especialidadSeleccionada.id, 'ocupado', false).subscribe(x => {
-      this.turnos = x;
+
+      let hoy = new Date();
+      let maximo = new Date();
+      maximo.setDate(hoy.getDate() + 15);
+
+      let turnosFiltradosPorFecha = x.filter((x: any) => {
+        let fechaDate = new Date(x.fechaDate.seconds * 1000);
+        if (fechaDate >= hoy && fechaDate <= maximo)
+          return true;
+        return false;
+      })
+
+      this.turnos = turnosFiltradosPorFecha;
       this.spinnerTurnos = false;
-      if (this.turnos.length == 0)
+      if (this.turnos.length == 0 && !this.cargando)
         this.toastService.informacion('El especialista seleccionado NO TIENE turnos disponibles para la especialidad seleccionada.', 'Aviso.');
     });
   }
@@ -86,26 +120,43 @@ export class SolicitarTurnoComponent {
   }
 
   seleccionEspecialidad(especialidad: any) {
-    this.especialistas = null;
-    this.turnos = null;
-    this.spinnerEspecialistas = true;
-    this.especialidadSeleccionada = especialidad;
-    this.eliminarSuscripcionTurnos();
-    this.suscripcionEspecialistas = this.usuarioService.obtenerEspecialistaPorEspecialidad(especialidad).subscribe(x => {
-      this.especialistas = x;
-      if (this.especialistas.length == 0)
-        this.toastService.informacion('La especialidad seleccionada NO CUENTA con especialistas por el momento.', 'Aviso.');
-      this.spinnerEspecialistas = false;
-    });
+    if (this.esAdmin() && this.pacienteSeleccionado == null) {
+      this.swalService.warning('Debe seleccionar un paciente para operar.', 'Aviso.');
+    } else {
+      this.filaSeleccionada = especialidad;
+      this.especialistas = null;
+      this.turnos = null;
+      this.spinnerEspecialistas = true;
+      this.especialidadSeleccionada = especialidad;
+      this.eliminarSuscripcionTurnos();
+      this.eliminarSuscripcionEspecialistas();
+      this.suscripcionEspecialistas = this.usuarioService.obtenerEspecialistaPorEspecialidad(especialidad).subscribe(x => {
+        this.especialistas = x;
+        if (this.especialistas.length == 0)
+          this.toastService.informacion('La especialidad seleccionada NO CUENTA con especialistas por el momento.', 'Aviso.');
+        this.spinnerEspecialistas = false;
+      });
+    }
   }
 
   seleccionTurno(turno: any) {
-    this.eliminarSuscripcionTurnos();
     this.turnos = null;
     this.cargando = true;
     turno.ocupado = true;
-    turno.idPaciente = this.usuarioActual.id;
-    turno.nombrePaciente = this.usuarioActual.apellido + ' ' + this.usuarioActual.nombre;
+    turno.estadoTurno = 'Solicitado';
+
+    let paciente;
+
+    if (this.esAdmin()) {
+      paciente = this.pacienteSeleccionado;
+    } else {
+      paciente = this.usuarioActual;
+    }
+
+    turno.paciente = paciente;
+    turno.idPaciente = paciente.id;
+    turno.nombrePaciente = paciente.apellido + ' ' + paciente.nombre;
+
     this.horariosService.modificarItem(turno).then(x => {
       setTimeout(() => {
         this.swalService.exito('Turno cargado con exito!');
@@ -113,4 +164,37 @@ export class SolicitarTurnoComponent {
       }, 1000);
     });
   }
+
+
+  ////////////////////-------------BUSCADOR----------------//////////////////////////////////////////
+  //LO DECLARO DE ESTA FORMA PARA QUE EL COMPONENTE 
+  //QUE LLAME A ESTA FUNCION PUEDA ACCEDER A LAS VARIABLES DE ESTE COMPONENTE
+  buscarPaciente = async (dato: any) => {
+    let usuarios = await this.usuarioService.traerListaFiltradaPor_UN_Campo('perfil', 1);
+    usuarios = usuarios.filter((x: any) => x?.nombre?.toLowerCase().includes(dato) || x?.apellido?.toLowerCase().includes(dato) || x?.dni?.toString().includes(dato));
+
+    usuarios = this.ordenarPorNombreYApellidoPipe.transform(usuarios);
+
+    return usuarios;
+  }
+
+  transformarTexto = (item: any) => {
+    return this.usuarioPipe.transform(item);
+  }
+
+  limpiarDatos() {
+    this.especialidadSeleccionada = null;
+    this.especialistas = null;
+    this.especialidades = null;
+    this.turnos = null;
+    this.pacienteSeleccionado = null;
+    this.filaSeleccionada = null;
+    this.traerEspecialidades();
+  }
+
+  obtenerUsuario(usuario: any) {
+    this.pacienteSeleccionado = usuario;
+  }
+  ////////////////////-------------FIN BUSCADOR----------------//////////////////////////////////////////
+
 }
